v0.1.12: fix INTEGER→i64 panics + e2e test script bug
Backend:
- Fix sqlx panics: Postgres INTEGER columns read as Rust i64 causes
  runtime type mismatch. Added ::BIGINT casts in all affected queries:
  - feed.rs: getNotes (version), getGlobalArchiveFeed (session_count)
  - search.rs: searchNotes (version), searchArchive (session_count)
  - note.rs: versionNote, getNoteHistory, getCollectiveNote,
    acceptEdit (version in notes + collective_notes)
  - archive.rs: sealArchive, getArchive (session_count, note_count)

Test script:
- Fix run_test/run_test_body curl exit code capture. The pattern
  $(curl ... || echo "000") concatenates curl's stdout with the
  fallback, producing '000000'. Changed to:
  http_code=$(curl ...) || http_code="000"

Root cause of the getNotes e2e failure: the server panicked on
r.get::<i64, _>("version") because the notes.version column is
Postgres INTEGER (int4/i32), but sqlx's Decode<i64> expects BIGINT
(int8). The ::BIGINT cast in SQL makes Postgres return int8, matching
the Rust type.

+++ src/handlers/archive.rs
@@ -134,7 +134,7 @@ pub async fn seal_archive(
 
     // Validate the archive exists with status 'initiated'.
     let archive = sqlx::query_as::<_, (i64, i64)>(
-        "SELECT session_count, note_count FROM archives \
+        "SELECT session_count::BIGINT, note_count::BIGINT FROM archives \
          WHERE course_uri = $1 AND semester = $2 AND status = 'initiated'",
     )
     .bind(&input.course_uri)
@@ -368,7 +368,7 @@ pub async fn get_archive(
         ),
     >(
         "SELECT archive_uri, course_uri, semester, sealed_by, ring_did, cid, \
-                internet_archive_url, session_count, note_count, status, \
+                internet_archive_url, session_count::BIGINT, note_count::BIGINT, status, \
                 initiated_at, sealed_at \
          FROM archives \
          WHERE course_uri = $1 AND semester = $2",
